move ProjectConfig type to common types package
