+++ packages/core/src/builder.ts
@@ -23,10 +23,11 @@ import {
   FeatureKey,
   Allocation,
   EnvironmentKey,
+  ProjectConfig,
 } from "@featurevisor/types";
 import { MAX_BUCKETED_NUMBER } from "@featurevisor/sdk";
 
-import { ProjectConfig, SCHEMA_VERSION } from "./config";
+import { SCHEMA_VERSION } from "./config";
 import { getNewTraffic } from "./traffic";
 import {
   parseYaml,
